Refactor: consolidate DSpark decode output projection
- Keep decode parallel configuration limited to TP, DP, and EP
- Consolidate attention collectives and TP-sharded output projection
  in decode_o_proj.py
