+++ models/deepseek_v4_flash_dspark/config.py
@@ -268,18 +268,9 @@ INT8_AMAX_EPS = 1e-4                      # amax floor: avoids 127/0 on all-zero
 FP32_NEG_INF = -3.4028234663852886e38     # most-negative finite fp32 (softmax masking)
 
 # Parallelism constants
-TP = 4         # tensor-parallel ranks per DP group
-SP = TP        # sequence-parallel token owners in the TP group
-DP = 4         # DP groups per node
-EP = 16        # expert-parallel world size (moe overrides it from --ep)
-
-# Per-component TP degree, over the components that shard.
-TP_Q_B = 1            # wq_b: replicated across the DSA-CP group
-TP_O_A = TP           # wo_a: ColumnParallel over o_groups
-TP_O_B = TP           # wo_b: RowParallel, then reduce-scatter
-TP_ATTN_SINK = 1      # attn_sink: all attention heads on every DSA-CP rank
-TP_SHARED_EXPERT = 1  # shared expert: sequence-parallel with replicated weights
-TP_VOCAB = TP         # embed_tokens / lm_head: vocab-parallel
+TP = 4    # tensor-parallel ranks per DP group
+DP = 4    # DP groups per node
+EP = 16   # expert-parallel world size (moe overrides it from --ep)
 
 # MoE constants
 MOE_TOKENS = DECODE_TOKENS * DP // EP
